✨ feat(origin-validation): move originMatchesAllowedEntry function to utils
